requirements.txt
streamlit
sentence-transformers
scikit-learn
python-docx
pdfminer.six
docx2txt